+++ src/test/java/ch/unibas/medizin/dynamicreports/test/jasper/style/TemplateStyle1Test.java
@@ -192,7 +192,7 @@ public class TemplateStyle1Test extends AbstractJasperStyleTest implements Seria
 
     @Override
     public Boolean evaluate(ReportParameters reportParameters) {
-      final Integer value = (Integer) reportParameters.getValue("field1");
+      final Integer value = reportParameters.getValue("field1");
       return values.contains(value);
     }
   }
